(PUP-7327) Use DataDigFunctionProvider to dispatch hiera3_backend calls
A Hiera version 3 backend must be responsible for digging into data when
dotted keys are used. That was not the case prior to this commit. Now,
the `DataDigFunctionProvider` will act as the dispatcher for the
`hiera3_backend` type instead of the `LookupKeyFunctionProvider`.
