+++ lib/puppet/pops/lookup/data_dig_function_provider.rb
@@ -14,41 +14,120 @@ class DataDigFunctionProvider < FunctionProvider
   # @return [Object] the found object
   # @throw :no_such_key when the object is not found
   def unchecked_key_lookup(key, lookup_invocation, merge)
-    segments = key.to_a
     lookup_invocation.with(:data_provider, self) do
       MergeStrategy.strategy(merge).lookup(locations, lookup_invocation) do |location|
-        if location.nil?
-          value = data_dig(key, segments, lookup_invocation, nil)
-          lookup_invocation.report_found(key, validate_data_value(self, value))
-          key.undig(value)
-        else
-          lookup_invocation.with(:location, location) do
-            if location.exist?
-              value = data_dig(key, segments, lookup_invocation, location.location)
-              lookup_invocation.report_found(key, validate_data_value(self, value))
-              key.undig(value)
-            else
-              lookup_invocation.report_path_not_found
-              throw :no_such_key
-            end
-          end
-        end
+        invoke_with_location(lookup_invocation, location, key, merge)
       end
     end
   end
 
+  def invoke_with_location(lookup_invocation, location, key, merge)
+    if location.nil?
+      value = data_dig(key, lookup_invocation, nil, merge)
+      lookup_invocation.report_found(key, validate_data_value(self, value))
+      key.undig(value)
+    else
+      lookup_invocation.with(:location, location) do
+        value = data_dig(key, lookup_invocation, location, merge)
+        lookup_invocation.report_found(key, validate_data_value(self, value))
+        key.undig(value)
+      end
+    end
+  end
+
+  def label
+    'Data Dig'
+  end
+
   private
 
-  def data_dig(key, segments, lookup_invocation, location)
+  def data_dig(key, lookup_invocation, location, merge)
+    unless location.nil? || location.exist?
+      lookup_invocation.report_location_not_found
+      throw :no_such_key
+    end
     ctx = function_context(lookup_invocation, location)
     ctx.data_hash ||= {}
     catch(:no_such_key) do
       hash = ctx.data_hash
-      hash[key] = ctx.function.call(lookup_invocation.scope, segments, options(location), Context.new(ctx, lookup_invocation)) unless hash.include?(key)
+      hash[key] = ctx.function.call(lookup_invocation.scope, key.to_a, options(location), Context.new(ctx, lookup_invocation)) unless hash.include?(key)
       return hash[key]
     end
     lookup_invocation.report_not_found(key)
     throw :no_such_key
+  end
+end
+
+# @api private
+class V3BackendFunctionProvider < DataDigFunctionProvider
+  TAG = 'hiera3_backend'.freeze
+
+  def data_dig(key, lookup_invocation, location, merge)
+    @backend ||= instantiate_backend(lookup_invocation)
+
+    # A merge_behavior retrieved from hiera.yaml must not be converted here. Instead, passing the symbol :hash
+    # tells the V3 backend to pick it up from the config.
+    resolution_type = lookup_invocation.hiera_v3_merge_behavior? ? :hash : convert_merge(merge)
+    @backend.lookup(key.to_s, lookup_invocation.scope, lookup_invocation.hiera_v3_location_overrides, resolution_type, context = {:recurse_guard => nil})
+  end
+
+  private
+
+  def instantiate_backend(lookup_invocation)
+    backend_name = options[HieraConfig::KEY_BACKEND]
+    begin
+      require 'hiera/backend'
+      require "hiera/backend/#{backend_name.downcase}_backend"
+      backend = Hiera::Backend.const_get("#{backend_name.capitalize}_backend").new
+      return backend.method(:lookup).arity == 4 ? Hiera::Backend::Backend1xWrapper.new(backend) : backend
+    rescue LoadError => e
+      lookup_invocation.report_text { "Unable to load backend '#{backend_name}': #{e.message}" }
+      throw :no_such_key
+    rescue NameError => e
+      lookup_invocation.report_text { "Unable to instantiate backend '#{backend_name}': #{e.message}" }
+      throw :no_such_key
+    end
+  end
+
+  # Converts a lookup 'merge' parameter argument into a Hiera 'resolution_type' argument.
+  #
+  # @param merge [String,Hash,nil] The lookup 'merge' argument
+  # @return [Symbol,Hash,nil] The Hiera 'resolution_type'
+  def convert_merge(merge)
+    case merge
+    when nil
+    when 'first', 'default'
+      # Nil is OK. Defaults to Hiera :priority
+      nil
+    when Puppet::Pops::MergeStrategy
+      convert_merge(merge.configuration)
+    when 'unique'
+      # Equivalent to Hiera :array
+      :array
+    when 'hash'
+      # Equivalent to Hiera :hash with default :native merge behavior. A Hash must be passed here
+      # to override possible Hiera deep merge config settings.
+      { :behavior => :native }
+    when 'deep', 'unconstrained_deep'
+      # Equivalent to Hiera :hash with :deeper merge behavior.
+      { :behavior => :deeper }
+    when 'reverse_deep'
+      # Equivalent to Hiera :hash with :deep merge behavior.
+      { :behavior => :deep }
+    when Hash
+      strategy = merge['strategy']
+      case strategy
+      when 'deep', 'unconstrained_deep', 'reverse_deep'
+        result = { :behavior => strategy == 'reverse_deep' ? :deep : :deeper }
+        # Remaining entries must have symbolic keys
+        merge.each_pair { |k,v| result[k.to_sym] = v unless k == 'strategy' }
+        result
+      else
+        convert_merge(strategy)
+      end
+    else
+      raise Puppet::DataBinding::LookupError, "Unrecognized value for request 'merge' parameter: '#{merge}'"
+    end
   end
 end
 end
